+++ docs/change-log.md
@@ -1,8 +1,13 @@
 # Change Log
 
 ---
-## 8 Sep 2017: V 1.8.2 (by FORTH-ICS)
+## 11 Sep 2017: V 1.8.2 (by FORTH-ICS)
 * Implemented new MERGE functionality ([issue #97](https://github.com/isl/x3ml/issues/97))
+* Updated X3ML Engine to support termonilogy functionalities (e.g. broader and exactMatch terms) ([issue #98](https://github.com/isl/x3ml/issues/98))
+* Updated X3ML schema to support termonilogy functionalities (e.g. broader and exactMatch terms) ([issue #99](https://github.com/isl/x3ml/issues/99))
+* Updated X3MLEngine executor to support transformations using SKOS terminologies ([issue #100](https://github.com/isl/x3ml/issues/100))
+* Updated X3MLEngineFactory SKOS terminologies ([issue #101](https://github.com/isl/x3ml/issues/101))
+
 
 ## 3 Aug 2017: V 1.8.1 (by FORTH-ICS)
 * Implemented new Generator RemoveTerm ([issue #95](https://github.com/isl/x3ml/issues/95))